+++ Minggu 5 - 13 Maret/minggu5/Nilai/Nilai.java
@@ -44,13 +44,14 @@ public class Nilai {
         }
     }
 
-    int rata(int[] arr) {
+    double rata(int[] arr) {
         int total = 0;
         for (int i = 0; i < arr.length; i++) {
             total += arr[i];
         }
-        return total / arr.length;
+        return (double) total / arr.length;
     }
 
+    
     
 }
